+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-2024-10-27 v1.14.27
+2024-10-27 v1.14.31
 
 - Fix: The executable remains open when there is an exception in the main app.
 - Fix: Avatar voice used when the voice was not enabled, and the avatar window was hidden.
@@ -6,6 +6,7 @@
 - New: Added the `ᵀ` suffix to temporary message attachments that are sent once.
 - New: [Options] tab for messages to control additional options when editing.
 - Updated: The buttons will now change to Apply/Cancel when editing a message.
+- Updated: Regenerate the chat from any user message.
 
 - 2024-10-25 v1.14.17